+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * fix async responsive greetings
 * fix multiple echo of async functions and promises mixed with strings
 * fix `get_output()` on lines that are promises [#687](https://github.com/jcubic/jquery.terminal/issues/687)
+* fix echo newline when one of the values is promise or function
 
 ## 2.29.1
 ### Bugfix